+++ src/main/java/eight/exception/NoSuchGenreException.java
@@ -1,4 +1,4 @@
-package rybka.exception;
+package eight.exception;
 
 public class NoSuchGenreException extends Exception {
     public NoSuchGenreException(String message) {
