+++ src/main/java/eight/exception/NoSuchTitleException.java
@@ -1,4 +1,4 @@
-package rybka.exception;
+package eight.exception;
 
 public class NoSuchTitleException extends Exception {
     public NoSuchTitleException(String message) {
