+++ src/main/java/eight/exception/NotAYearException.java
@@ -1,4 +1,4 @@
-package rybka.exception;
+package eight.exception;
 
 public class NotAYearException extends Exception {
     public NotAYearException(String message) {
