+++ src/main/java/eight/model/Movie.java
@@ -1,4 +1,4 @@
-package rybka.model;
+package eight.model;
 
 import java.time.LocalTime;
